src/test/kotlin/org/jglynn/aoc/Day03Test.kt
package org.jglynn.aoc

import org.assertj.core.api.Assertions.assertThat
import org.junit.jupiter.api.DisplayName
import org.junit.jupiter.api.Nested
import org.junit.jupiter.api.Test

@DisplayName("Day 3")
class Day03Test {

    // Arrange
    private val input = """
        00100
        11110
        10110
        10111
        10101
        01111
        00111
        11100
        10000
        11001
        00010
        01010
""".trimIndent().lines()

    @Nested
    @DisplayName("Part 1")
    inner class Part1 {
        @Test
        fun `Matches example`() {
            // Act
            val answer = Day03(input).solvePart1()

            // Assert
            assertThat(answer).isEqualTo(198)
        }

        @Test
        fun `Actual answer`() {
            // Act
            val answer = Day03(Utils.resourceAsListOfString("Day03_test.txt")).solvePart1()

            // Assert
            assertThat(answer).isEqualTo(845_186)
        }
    }
}
